move pow to arithmetic.py
